+++ packages/core/src/plugins/withPlate.ts
@@ -30,6 +30,7 @@ export const withPlate = <
 
   editor.id = id as string;
   editor.prevSelection = null;
+  editor.currentKeyboardEvent = null;
 
   if (!editor.key) {
     editor.key = Math.random();