+++ packages/nodes/table/src/transforms/index.ts
@@ -9,4 +9,5 @@ export * from './insertTable';
 export * from './insertTableColumn';
 export * from './insertTableRow';
 export * from './moveSelectionFromCell';
+export * from './overrideSelectionFromCell';
 export * from './setTableColSize';