+++ packages/nodes/table/src/transforms/moveSelectionFromCell.ts
@@ -9,6 +9,7 @@ import {
   Value,
   withoutNormalizing,
 } from '@udecode/plate-core';
+import { Location } from 'slate';
 import { getTableGridAbove } from '../queries/getTableGridAbove';
 import { getCellTypes } from '../utils/getCellType';
 
@@ -18,9 +19,13 @@ import { getCellTypes } from '../utils/getCellType';
 export const moveSelectionFromCell = <V extends Value = Value>(
   editor: PlateEditor<V>,
   {
+    at,
     reverse,
     edge,
+    fromOneCell,
   }: {
+    at?: Location;
+
     /**
      * false: move selection to cell below
      * true: move selection to cell above
@@ -30,12 +35,19 @@ export const moveSelectionFromCell = <V extends Value = Value>(
      * Expand cell selection to an edge.
      */
     edge?: 'top' | 'left' | 'right' | 'bottom';
+
+    /**
+     * Move selection from one selected cell
+     */
+    fromOneCell?: boolean;
   } = {}
 ) => {
   if (edge) {
-    const cellEntries = getTableGridAbove(editor, { format: 'cell' });
+    const cellEntries = getTableGridAbove(editor, { at, format: 'cell' });
 
-    if (cellEntries.length > 0) {
+    const minCell = fromOneCell ? 0 : 1;
+
+    if (cellEntries.length > minCell) {
       const [, firstCellPath] = cellEntries[0];
       const [, lastCellPath] = cellEntries[cellEntries.length - 1];
 
@@ -64,6 +76,7 @@ export const moveSelectionFromCell = <V extends Value = Value>(
   }
 
   const cellEntry = getBlockAbove(editor, {
+    at,
     match: { type: getCellTypes(editor) },
   });