+++ packages/ui/nodes/table/src/TableElement/TableElement.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { Value, withProviders } from '@udecode/plate-core';
+import { Value } from '@udecode/plate-core';
 import { getRootProps } from '@udecode/plate-styled-components';
-import { ELEMENT_TABLE } from '@udecode/plate-table';
-import { Provider } from 'jotai';
 import { useSelectedCells } from '../hooks/useSelectedCells';
 import { useTableColSizes } from '../hooks/useTableColSizes';
 import { useTableStore } from '../table.atoms';
@@ -10,9 +8,9 @@ import { TablePopover } from '../TablePopover/TablePopover';
 import { getTableElementStyles } from './TableElement.styles';
 import { TableElementProps } from './TableElement.types';
 
-export const TableElementBase = <V extends Value>({
+export const TableElement = <V extends Value>({
   transformColSizes,
-  floatingOptions,
+  popoverProps,
   ...props
 }: TableElementProps<V>) => {
   const { attributes, children, nodeProps, element } = props;
@@ -48,7 +46,7 @@ export const TableElementBase = <V extends Value>({
         ))}
       </colgroup>
 
-      <TablePopover floatingOptions={floatingOptions}>
+      <TablePopover {...popoverProps}>
         <tbody css={tbody?.css} className={tbody?.className}>
           {children}
         </tbody>
@@ -56,7 +54,3 @@ export const TableElementBase = <V extends Value>({
     </table>
   );
 };
-
-export const TableElement = withProviders([Provider, { scope: ELEMENT_TABLE }])(
-  TableElementBase
-);