+++ packages/nodes/horizontal-rule/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-horizontal-rule
 
+## 16.5.0
+
+## 16.3.0
+
 ## 16.2.0
 
 ## 16.1.0