+++ packages/ui/nodes/table/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @udecode/plate-table-ui
 
+## 16.6.0
+
+## 16.5.0
+
+## 16.4.2
+
+### Patch Changes
+
+- [#1821](https://github.com/udecode/plate/pull/1821) by [@zbeyens](https://github.com/zbeyens) –
+  - Fixes https://github.com/udecode/editor-protocol/issues/77
+  - Fixes #1820
+
+## 16.4.0
+
+### Minor Changes
+
+- [#1800](https://github.com/udecode/plate/pull/1800) by [@zbeyens](https://github.com/zbeyens) – Light breaking changes:
+  - `TableElementBase` has been removed for `TableElement`
+  - in v16, `TableElement` `onRenderContainer` prop has been removed for `floatingOptions`. However, `TablePopover` has other props which you may want to control. So this release adds `popoverProps` prop has been added and `floatingOptions` prop is moved into `popoverProps.floatingOptions`.
+
+## 16.3.0
+
 ## 16.2.0
 
 ### Patch Changes